requirements.txt
Django==4.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
Pillow==8.4.0